+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Use `<button>` elements for pagination if no `path` is set.
 - Mark up disabled previous/next links of the pagination as
   `<a role="link" aria-disabled="true">` instead of `<span>`.
 - Update documentation for `hidden_inputs_for_filter/1` to use